Refactor package
api 패키지와 ssr(타임리프) 패키지 분리
