+++ src/main/java/com/seungki/urlshortener/ssr/domain/UrlMapping.java
@@ -1,4 +1,4 @@
-package com.seungki.urlshortener.domain;
+package com.seungki.urlshortener.ssr.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
