+++ src/test/java/com/seungki/urlshortener/service/UrlShortenerServiceTest.java
@@ -5,8 +5,9 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-import com.seungki.urlshortener.domain.UrlMapping;
-import com.seungki.urlshortener.repository.UrlMappingRepository;
+import com.seungki.urlshortener.ssr.domain.UrlMapping;
+import com.seungki.urlshortener.ssr.repository.UrlMappingRepository;
+import com.seungki.urlshortener.ssr.service.UrlShortenerService;
 import jakarta.persistence.EntityManager;
 import java.time.LocalDateTime;
 import java.time.temporal.ChronoUnit;
